only copy when volume is empty

+++ Docker/SolrServer/docker-entrypoint-initdb.d/0_SOLR_HOME.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-if [ "${SOLR_HOME}" != "/var/solr/data" ]; then
-    cp -rv /var/solr/data/* "${SOLR_HOME}/"
+if [ "${SOLR_HOME}" != "/var/solr/data" ] && [ ! -f "${SOLR_HOME}/solr.xml" ]; then
+  cp -rv /var/solr/data/* "${SOLR_HOME}/"
 fi
